Add benchmarking check in pr (#113)
Add benchmarking check that compares the `pr` with the `main` branch.  
Check fails if the `pr` is determined to be meaningfully slower than `main`.  

The definition of "meaningfully slower" is arbitrary and may be changed based on experience.

Comparison is done using the tool `ResultsComparer`.  
The tool is copied from the https://github.com/dotnet/performance repo.  
https://github.com/dotnet/performance/blob/9f6dc6612a125da093c01dcdca47ad8573a53d0c/src/tools/ResultsComparer/README.md